refactor: nested endpoint directories for configuration: policies
