+++ vmngclient/endpoints/configuration/policy/vsmart_template.py
