+++ vmngclient/endpoints/endpoints_container.py
@@ -6,26 +6,26 @@ from vmngclient.endpoints.administration_user_and_group import AdministrationUse
 from vmngclient.endpoints.certificate_management_vmanage import CertificateManagementVManage
 from vmngclient.endpoints.client import Client
 from vmngclient.endpoints.cluster_management import ClusterManagement
+from vmngclient.endpoints.configuration.policy.definition_builder.data import ConfigurationPolicyDataDefinitionBuilder
+from vmngclient.endpoints.configuration.policy.definition_builder.zone_based_firewall import (
+    ConfigurationPolicyZoneBasedFirewallDefinitionBuilder,
+)
+from vmngclient.endpoints.configuration.policy.list_builder.data_prefix import ConfigurationPolicyDataPrefixListBuilder
+from vmngclient.endpoints.configuration.policy.list_builder.fqdn import ConfigurationPolicyFQDNListBuilder
+from vmngclient.endpoints.configuration.policy.list_builder.geo_location import (
+    ConfigurationPolicyGeoLocationListBuilder,
+)
+from vmngclient.endpoints.configuration.policy.list_builder.site_list_builder import ConfigurationPolicySiteListBuilder
+from vmngclient.endpoints.configuration.policy.list_builder.vpn import ConfigurationPolicyVPNListBuilder
+from vmngclient.endpoints.configuration.policy.list_builder.zone import ConfigurationPolicyZoneListBuilder
+from vmngclient.endpoints.configuration.policy.vsmart_template import ConfigurationVSmartTemplatePolicy
 from vmngclient.endpoints.configuration_dashboard_status import ConfigurationDashboardStatus
 from vmngclient.endpoints.configuration_device_actions import ConfigurationDeviceActions
 from vmngclient.endpoints.configuration_device_software_update import ConfigurationDeviceSoftwareUpdate
 from vmngclient.endpoints.configuration_device_template import ConfigurationDeviceTemplate
 from vmngclient.endpoints.configuration_feature_profile import ConfigurationFeatureProfile
 from vmngclient.endpoints.configuration_group import ConfigurationGroup
-from vmngclient.endpoints.configuration_policy_data_definition_builder import ConfigurationPolicyDataDefinitionBuilder
-from vmngclient.endpoints.configuration_policy_data_prefix_list_builder import ConfigurationPolicyDataPrefixListBuilder
-from vmngclient.endpoints.configuration_policy_fqdn_list_builder import ConfigurationPolicyFQDNListBuilder
-from vmngclient.endpoints.configuration_policy_geo_location_list_builder import (
-    ConfigurationPolicyGeoLocationListBuilder,
-)
-from vmngclient.endpoints.configuration_policy_site_list_builder import ConfigurationPolicySiteListBuilder
-from vmngclient.endpoints.configuration_policy_vpn_list_builder import ConfigurationPolicyVPNListBuilder
-from vmngclient.endpoints.configuration_policy_zone_based_firewall_definition_builder import (
-    ConfigurationPolicyZoneBasedFirewallDefinitionBuilder,
-)
-from vmngclient.endpoints.configuration_policy_zone_list_builder import ConfigurationPolicyZoneListBuilder
 from vmngclient.endpoints.configuration_settings import ConfigurationSettings
-from vmngclient.endpoints.configuration_vsmart_template_policy import ConfigurationVSmartTemplatePolicy
 from vmngclient.endpoints.monitoring_device_details import MonitoringDeviceDetails
 from vmngclient.endpoints.monitoring_status import MonitoringStatus
 from vmngclient.endpoints.sdavc_cloud_connector import SDAVCCloudConnector
